DMA UART works

+++ MCU_2_Navigation/MCU_2_Nav.X/DMA_Common.h
@@ -140,7 +140,7 @@ typedef struct
     uint8_t         channelPriority;
     _Bool           interruptEnabled;
     _Bool           initialized;
-    void            (*InterruptTriggerFnc)(uint32_t*); // event
+    void            (*InterruptTriggerFnc)(uint8_t*); // event
 } DMA;
